+++ app/src/main/java/org/oxycblt/auxio/detail/recycler/DetailAdapter.kt
@@ -28,6 +28,7 @@ import org.oxycblt.auxio.detail.SortHeader
 import org.oxycblt.auxio.list.Header
 import org.oxycblt.auxio.list.Item
 import org.oxycblt.auxio.list.SelectableListListener
+import org.oxycblt.auxio.list.UpdateInstructions
 import org.oxycblt.auxio.list.recycler.*
 import org.oxycblt.auxio.music.Music
 import org.oxycblt.auxio.util.context
@@ -44,7 +45,8 @@ abstract class DetailAdapter(
     private val listener: Listener<*>,
     diffCallback: DiffUtil.ItemCallback<Item>
 ) :
-    SelectionIndicatorAdapter<Item, RecyclerView.ViewHolder>(ListDiffer.Async(diffCallback)),
+    SelectionIndicatorAdapter<Item, UpdateInstructions, RecyclerView.ViewHolder>(
+        ListDiffer.Async(diffCallback)),
     AuxioRecyclerView.SpanSizeLookup {
 
     override fun getItemViewType(position: Int) =