+++ app/src/main/java/org/oxycblt/auxio/home/list/SongListFragment.kt
@@ -160,7 +160,7 @@ class SongListFragment :
      * @param listener An [SelectableListListener] to bind interactions to.
      */
     private class SongAdapter(private val listener: SelectableListListener<Song>) :
-        SelectionIndicatorAdapter<Song, SongViewHolder>(
+        SelectionIndicatorAdapter<Song, UpdateInstructions, SongViewHolder>(
             ListDiffer.Async(SongViewHolder.DIFF_CALLBACK)) {
 
         override fun onCreateViewHolder(parent: ViewGroup, viewType: Int) =