+++ app/src/main/java/org/oxycblt/auxio/search/SearchAdapter.kt
@@ -29,7 +29,8 @@ import org.oxycblt.auxio.music.*
  * @author Alexander Capehart (OxygenCobalt)
  */
 class SearchAdapter(private val listener: SelectableListListener<Music>) :
-    SelectionIndicatorAdapter<Item, RecyclerView.ViewHolder>(ListDiffer.Async(DIFF_CALLBACK)),
+    SelectionIndicatorAdapter<Item, UpdateInstructions, RecyclerView.ViewHolder>(
+        ListDiffer.Async(DIFF_CALLBACK)),
     AuxioRecyclerView.SpanSizeLookup {
 
     override fun getItemViewType(position: Int) =